backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
tortoise-orm[aiomysql]>=0.19.0
aerich>=0.7.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
imapclient==3.0.1
aiosmtplib==3.0.1
email-validator==2.1.0
python-dotenv==1.0.0
cryptography==41.0.7
loguru>=0.7.0
